+++ tests_basic/test_planar_imaging.py
@@ -10,25 +10,34 @@ from scipy.ndimage import rotate
 
 from pylinac import LeedsTOR, StandardImagingQC3, LasVegas, DoselabMC2kV, DoselabMC2MV
 from pylinac.core import image
-from pylinac.planar_imaging import PlanarResult, SNCkV, SNCMV, StandardImagingQCkV, PTWEPIDQC, StandardImagingFC2, \
-    IMTLRad, SNCFSQA, LeedsTORBlue
+from pylinac.planar_imaging import (
+    PlanarResult,
+    SNCkV,
+    SNCMV,
+    StandardImagingQCkV,
+    PTWEPIDQC,
+    StandardImagingFC2,
+    IMTLRad,
+    SNCFSQA,
+    LeedsTORBlue,
+    SNCMV12510,
+)
 from tests_basic.utils import save_file, CloudFileMixin, get_file_from_cloud_test_repo
 
-TEST_DIR = 'planar_imaging'
+TEST_DIR = "planar_imaging"
 
 
 class GeneralTests(TestCase):
-
     def test_from_file_object(self):
-        path = get_file_from_cloud_test_repo([TEST_DIR, 'Leeds_ccw.dcm'])
-        with open(path, 'rb') as f:
+        path = get_file_from_cloud_test_repo([TEST_DIR, "Leeds_ccw.dcm"])
+        with open(path, "rb") as f:
             phan = LeedsTOR(f)
             phan.analyze()
         self.assertIsInstance(phan, LeedsTOR)
 
     def test_from_stream(self):
-        path = get_file_from_cloud_test_repo([TEST_DIR, 'Leeds_ccw.dcm'])
-        with open(path, 'rb') as f:
+        path = get_file_from_cloud_test_repo([TEST_DIR, "Leeds_ccw.dcm"])
+        with open(path, "rb") as f:
             s = io.BytesIO(f.read())
             phan = LeedsTOR(s)
             phan.analyze()
@@ -43,12 +52,15 @@ class GeneralTests(TestCase):
         phan.analyze()
         data = phan.results_data()
         self.assertIsInstance(data, PlanarResult)
-        self.assertEqual(data.median_contrast, np.median([roi.contrast for roi in phan.low_contrast_rois]))
+        self.assertEqual(
+            data.median_contrast,
+            np.median([roi.contrast for roi in phan.low_contrast_rois]),
+        )
 
         data_dict = phan.results_data(as_dict=True)
         self.assertIsInstance(data_dict, dict)
         self.assertEqual(len(data_dict), 8)
-        self.assertIn('pylinac_version', data_dict)
+        self.assertIn("pylinac_version", data_dict)
 
     def test_results_data_no_mtf(self):
         phan = LasVegas.from_demo_image()
@@ -77,16 +89,16 @@ class GeneralTests(TestCase):
         phan.analyze()
         figs, names = phan.plot_analyzed_image(split_plots=True)
         self.assertEqual(len(figs), 3)
-        files = phan.save_analyzed_image(filename='a.png', split_plots=True)
-        names = ('a_image.png', 'a_low_contrast.png', 'a_high_contrast.png')
+        files = phan.save_analyzed_image(filename="a.png", split_plots=True)
+        names = ("a_image.png", "a_low_contrast.png", "a_high_contrast.png")
         for name in names:
             self.assertIn(name, files)
 
         # regular single plot produces one image/file
         figs, names = phan.plot_analyzed_image()
         self.assertEqual(len(figs), 0)
-        name = 'b.png'
-        phan.save_analyzed_image('b.png')
+        name = "b.png"
+        phan.save_analyzed_image("b.png")
         self.assertTrue(osp.isfile(name))
 
         # stream buffer shouldn't fail
@@ -100,7 +112,7 @@ class GeneralTests(TestCase):
             phan.save_analyzed_image()  # no filename and no streams is an error
 
     def test_passing_image_kwargs(self):
-        path = get_file_from_cloud_test_repo([TEST_DIR, 'Leeds_ccw.dcm'])
+        path = get_file_from_cloud_test_repo([TEST_DIR, "Leeds_ccw.dcm"])
 
         # do normal analysis
         phan = LeedsTOR(path)
@@ -109,7 +121,7 @@ class GeneralTests(TestCase):
 
         # pass kwarg; use same dpi as image; results should be the same.
         img = image.load(path)
-        phan = LeedsTOR(path, image_kwargs={'dpi': img.dpi})
+        phan = LeedsTOR(path, image_kwargs={"dpi": img.dpi})
         phan.analyze()
         x_manual_dpi = phan.results_data().phantom_center_x_y[0]
 
@@ -118,7 +130,7 @@ class GeneralTests(TestCase):
 
 class PlanarPhantomMixin(CloudFileMixin):
     klass: Callable
-    dir_path = ['planar_imaging']
+    dir_path = ["planar_imaging"]
     mtf_50 = None
     invert = False
     ssd = 1000
@@ -142,13 +154,15 @@ class PlanarPhantomMixin(CloudFileMixin):
 
     @classmethod
     def tearDownClass(cls):
-        plt.close('all')
+        plt.close("all")
         del cls.instance
 
     def test_plotting(self):
         self.instance.plot_analyzed_image()
         self.instance.plot_analyzed_image(low_contrast=False, high_contrast=False)
-        self.instance.plot_analyzed_image(image=False, low_contrast=False, high_contrast=False)
+        self.instance.plot_analyzed_image(
+            image=False, low_contrast=False, high_contrast=False
+        )
 
     def test_saving(self):
         self.instance.plot_analyzed_image()
@@ -159,19 +173,29 @@ class PlanarPhantomMixin(CloudFileMixin):
 
     def test_mtf(self):
         if self.mtf_50 is not None:
-            self.assertAlmostEqual(self.mtf_50, self.instance.mtf.relative_resolution(50), delta=0.3)
+            self.assertAlmostEqual(
+                self.mtf_50, self.instance.mtf.relative_resolution(50), delta=0.3
+            )
 
     def test_rois_seen(self):
         if self.rois_seen is not None:
-            self.assertEqual(self.rois_seen, self.instance.results_data().num_contrast_rois_seen)
+            self.assertEqual(
+                self.rois_seen, self.instance.results_data().num_contrast_rois_seen
+            )
 
     def test_median_contrast(self):
         if self.median_contrast is not None:
-            self.assertAlmostEqual(self.median_contrast, self.instance.results_data().median_contrast, delta=0.03)
+            self.assertAlmostEqual(
+                self.median_contrast,
+                self.instance.results_data().median_contrast,
+                delta=0.03,
+            )
 
     def test_median_cnr(self):
         if self.median_cnr is not None:
-            self.assertAlmostEqual(self.median_cnr, self.instance.results_data().median_cnr, delta=0.1)
+            self.assertAlmostEqual(
+                self.median_cnr, self.instance.results_data().median_cnr, delta=0.1
+            )
 
     def test_results(self):
         self.assertIsInstance(self.instance.results(), str)
@@ -188,7 +212,7 @@ class LeedsDemo(PlanarPhantomMixin, TestCase):
 class LeedsCCW(PlanarPhantomMixin, TestCase):
     klass = LeedsTOR
     mtf_50 = 1.5
-    file_name = 'Leeds_ccw.dcm'
+    file_name = "Leeds_ccw.dcm"
 
 
 class Leeds45Deg(PlanarPhantomMixin, TestCase):
@@ -196,34 +220,36 @@ class Leeds45Deg(PlanarPhantomMixin, TestCase):
     invert = True  # inverted in v3.0 due to changed default inversion behavior
     mtf_50 = 1.9
     ssd = 1500
-    file_name = 'Leeds-45deg.dcm'
+    file_name = "Leeds-45deg.dcm"
 
 
 class LeedsDirtyEdges(PlanarPhantomMixin, TestCase):
     klass = LeedsTOR
     mtf_50 = 1.3
     ssd = 1000
-    file_name = 'Leeds-dirty-edges.dcm'
+    file_name = "Leeds-dirty-edges.dcm"
 
 
 class LeedsBlue(PlanarPhantomMixin, TestCase):
     klass = LeedsTORBlue
-    dir_path = ['planar_imaging', 'Leeds']
+    dir_path = ["planar_imaging", "Leeds"]
     mtf_50 = 1.5
     ssd = 1450
-    file_name = 'Leeds_Blue.dcm'
+    file_name = "Leeds_Blue.dcm"
 
 
 class LeedsBlueRotated(PlanarPhantomMixin, TestCase):
     klass = LeedsTORBlue
-    dir_path = ['planar_imaging', 'Leeds']
+    dir_path = ["planar_imaging", "Leeds"]
     mtf_50 = 1.5
     ssd = 1450
-    file_name = 'Leeds_Blue.dcm'
+    file_name = "Leeds_Blue.dcm"
 
     @classmethod
     def preprocess(cls, instance):
-        instance.image.array = rotate(instance.image.array, angle=180, mode='mirror', order=0)
+        instance.image.array = rotate(
+            instance.image.array, angle=180, mode="mirror", order=0
+        )
 
 
 @skip("Phantom appears distorted. MTF locations are different than other phantoms")
@@ -231,13 +257,13 @@ class LeedsClosedBlades(PlanarPhantomMixin, TestCase):
     klass = LeedsTOR
     mtf_50 = 1.3
     ssd = 1500
-    file_name = 'Leeds-closed-blades.dcm'
+    file_name = "Leeds-closed-blades.dcm"
 
 
 class LeedsACB1(PlanarPhantomMixin, TestCase):
     klass = LeedsTOR
-    dir_path = ['planar_imaging', 'ACB 1']
-    file_path = '1.dcm'
+    dir_path = ["planar_imaging", "ACB 1"]
+    file_path = "1.dcm"
     mtf_50 = 1.4
 
 
@@ -252,14 +278,14 @@ class SIQC3Demo(PlanarPhantomMixin, TestCase):
 
 class SIQC3_1(PlanarPhantomMixin, TestCase):
     klass = StandardImagingQC3
-    file_name = 'QC3-2.5MV.dcm'
+    file_name = "QC3-2.5MV.dcm"
     mtf_50 = 1.19
     rois_seen = 5
 
 
 class SIQC3_2(PlanarPhantomMixin, TestCase):
     klass = StandardImagingQC3
-    file_name = 'QC3-2.5MV-2.dcm'
+    file_name = "QC3-2.5MV-2.dcm"
     mtf_50 = 1.16
     rois_seen = 5
 
@@ -289,21 +315,36 @@ class LasVegasDemo(LasVegasTestMixin, TestCase):
         LasVegas.run_demo()  # shouldn't raise
 
 
-@skip("Non-cardinal angles no longer supported. If support is re-added these can be reactivated")
+@skip(
+    "Non-cardinal angles no longer supported. If support is re-added these can be reactivated"
+)
 class LasVegas10deg(LasVegasTestMixin, TestCase):
-    file_path = ['TrueBeam 1 - 2364', '2.5MV LV HQ 10deg - ImageRT_2016-10-6 20-12-58.dcm']
+    file_path = [
+        "TrueBeam 1 - 2364",
+        "2.5MV LV HQ 10deg - ImageRT_2016-10-6 20-12-58.dcm",
+    ]
     phantom_angle = 290
 
 
-@skip("Non-cardinal angles no longer supported. If support is re-added these can be reactivated")
+@skip(
+    "Non-cardinal angles no longer supported. If support is re-added these can be reactivated"
+)
 class LasVegasrotated(LasVegasTestMixin, TestCase):
-    file_path = ['TrueBeam 1 - 2364', '2.5MV LV HQ side2 - ImageRT_2016-10-6 20-43-3.dcm']
+    file_path = [
+        "TrueBeam 1 - 2364",
+        "2.5MV LV HQ side2 - ImageRT_2016-10-6 20-43-3.dcm",
+    ]
     phantom_angle = 284
 
 
-@skip("Non-cardinal angles no longer supported. If support is re-added these can be reactivated")
+@skip(
+    "Non-cardinal angles no longer supported. If support is re-added these can be reactivated"
+)
 class LasVegasTB1(LasVegasTestMixin, TestCase):
-    file_path = ['TrueBeam 1 - 2364', '6MV LasVegas HQ 0deg - ImageRT_2016-10-6 20-10-17.dcm']
+    file_path = [
+        "TrueBeam 1 - 2364",
+        "6MV LasVegas HQ 0deg - ImageRT_2016-10-6 20-10-17.dcm",
+    ]
     phantom_angle = 284.5
 
 
@@ -339,6 +380,23 @@ class SNCMVDemo(PlanarPhantomMixin, TestCase):
         SNCkV.run_demo()
 
 
+class SNCMV12510_6MV1(PlanarPhantomMixin, TestCase):
+    klass = SNCMV12510
+    mtf_50 = 0.91
+    dir_path = ["planar_imaging", "SNC MV Old"]
+    file_name = "SNC_MV_Old1.dcm"
+
+    def test_demo(self):
+        SNCMV12510.run_demo()
+
+
+class SNCMV12510_6MV2(PlanarPhantomMixin, TestCase):
+    klass = SNCMV12510
+    mtf_50 = 0.85
+    dir_path = ["planar_imaging", "SNC MV Old"]
+    file_name = "SNC_MV_Old2.dcm"
+
+
 class SIQCkVDemo(PlanarPhantomMixin, TestCase):
     klass = StandardImagingQCkV
     mtf_50 = 1.81
@@ -362,8 +420,8 @@ class PTWEPIDQC1(PlanarPhantomMixin, TestCase):
     rois_seen = 9
     median_contrast = 0.26
     median_cnr = 40.9
-    dir_path = ['planar_imaging', 'PTW-EPID']
-    file_name = 'PTW EPID QC Phantom.dcm'
+    dir_path = ["planar_imaging", "PTW-EPID"]
+    file_name = "PTW EPID QC Phantom.dcm"
 
 
 class PTWEPID15MV(PlanarPhantomMixin, TestCase):
@@ -372,8 +430,8 @@ class PTWEPID15MV(PlanarPhantomMixin, TestCase):
     rois_seen = 8
     median_contrast = 0.17
     median_cnr = 26.7
-    dir_path = ['planar_imaging', 'PTW-EPID']
-    file_name = 'PTW-EPID-15MV.dcm'
+    dir_path = ["planar_imaging", "PTW-EPID"]
+    file_name = "PTW-EPID-15MV.dcm"
 
 
 class PTWEPID6xHigh(PlanarPhantomMixin, TestCase):
@@ -382,8 +440,8 @@ class PTWEPID6xHigh(PlanarPhantomMixin, TestCase):
     rois_seen = 9
     median_contrast = 0.28
     median_cnr = 72.1
-    dir_path = ['planar_imaging', 'PTW-EPID']
-    file_name = 'QA EPI 6x High.dcm'
+    dir_path = ["planar_imaging", "PTW-EPID"]
+    file_name = "QA EPI 6x High.dcm"
 
 
 class PTWEPID6xHighQuality(PlanarPhantomMixin, TestCase):
@@ -392,8 +450,8 @@ class PTWEPID6xHighQuality(PlanarPhantomMixin, TestCase):
     rois_seen = 9
     median_contrast = 0.254
     median_cnr = 37.9
-    dir_path = ['planar_imaging', 'PTW-EPID']
-    file_name = 'TB1_PTW_EPID_Phan 6x_HighQuality.dcm'
+    dir_path = ["planar_imaging", "PTW-EPID"]
+    file_name = "TB1_PTW_EPID_Phan 6x_HighQuality.dcm"
 
 
 class PTWEPIDTB3(PlanarPhantomMixin, TestCase):
@@ -402,8 +460,8 @@ class PTWEPIDTB3(PlanarPhantomMixin, TestCase):
     rois_seen = 9
     median_contrast = 0.31
     median_cnr = 43.2
-    dir_path = ['planar_imaging', 'PTW-EPID']
-    file_name = 'TB3_PTW_EPID_Phan.dcm'
+    dir_path = ["planar_imaging", "PTW-EPID"]
+    file_name = "TB3_PTW_EPID_Phan.dcm"
 
 
 class PTWEPIDTB4(PlanarPhantomMixin, TestCase):
@@ -412,13 +470,13 @@ class PTWEPIDTB4(PlanarPhantomMixin, TestCase):
     rois_seen = 9
     median_contrast = 0.30
     median_cnr = 39.1
-    dir_path = ['planar_imaging', 'PTW-EPID']
-    file_name = 'TB4 PTW_EPID_phan.dcm'
+    dir_path = ["planar_imaging", "PTW-EPID"]
+    file_name = "TB4 PTW_EPID_phan.dcm"
 
 
 class FC2Mixin(PlanarPhantomMixin):
     klass = StandardImagingFC2
-    dir_path = ['planar_imaging', 'SI FC2']
+    dir_path = ["planar_imaging", "SI FC2"]
     field_size_x_mm = 150
     field_size_y_mm = 150
     field_epid_offset_x_mm = 0
@@ -440,12 +498,24 @@ class FC2Mixin(PlanarPhantomMixin):
 
     def test_field_size(self):
         results_data = self.instance.results_data()
-        assert results_data.field_size_x_mm == pytest.approx(self.field_size_x_mm, abs=0.3)
-        assert results_data.field_size_y_mm == pytest.approx(self.field_size_y_mm, abs=0.3)
-        assert results_data.field_epid_offset_x_mm == pytest.approx(self.field_epid_offset_x_mm, abs=0.2)
-        assert results_data.field_epid_offset_y_mm == pytest.approx(self.field_epid_offset_y_mm, abs=0.2)
-        assert results_data.field_bb_offset_x_mm == pytest.approx(self.field_bb_offset_x_mm, abs=0.2)
-        assert results_data.field_bb_offset_y_mm == pytest.approx(self.field_bb_offset_y_mm, abs=0.2)
+        assert results_data.field_size_x_mm == pytest.approx(
+            self.field_size_x_mm, abs=0.3
+        )
+        assert results_data.field_size_y_mm == pytest.approx(
+            self.field_size_y_mm, abs=0.3
+        )
+        assert results_data.field_epid_offset_x_mm == pytest.approx(
+            self.field_epid_offset_x_mm, abs=0.2
+        )
+        assert results_data.field_epid_offset_y_mm == pytest.approx(
+            self.field_epid_offset_y_mm, abs=0.2
+        )
+        assert results_data.field_bb_offset_x_mm == pytest.approx(
+            self.field_bb_offset_x_mm, abs=0.2
+        )
+        assert results_data.field_bb_offset_y_mm == pytest.approx(
+            self.field_bb_offset_y_mm, abs=0.2
+        )
 
 
 class FC2Demo(FC2Mixin, TestCase):
@@ -461,7 +531,7 @@ class FC2Demo(FC2Mixin, TestCase):
 
 
 class FC210x10_10FFF(FC2Mixin, TestCase):
-    file_name = 'FC-2-10x10-10fff.dcm'
+    file_name = "FC-2-10x10-10fff.dcm"
     field_size_y_mm = 98.7
     field_size_x_mm = 99.3
     field_epid_offset_x_mm = 0
@@ -471,7 +541,7 @@ class FC210x10_10FFF(FC2Mixin, TestCase):
 
 
 class FC210x10_10X(FC2Mixin, TestCase):
-    file_name = 'FC-2-10x10-10x.dcm'
+    file_name = "FC-2-10x10-10x.dcm"
     field_size_y_mm = 99.3
     field_size_x_mm = 99.6
     field_epid_offset_x_mm = 0.2
@@ -481,7 +551,7 @@ class FC210x10_10X(FC2Mixin, TestCase):
 
 
 class FC210x10_15X(FC2Mixin, TestCase):
-    file_name = 'FC-2-10x10-15x.dcm'
+    file_name = "FC-2-10x10-15x.dcm"
     field_size_y_mm = 99.3
     field_size_x_mm = 99.6
     field_epid_offset_x_mm = 0.1
@@ -491,7 +561,7 @@ class FC210x10_15X(FC2Mixin, TestCase):
 
 
 class FC215x15_10X(FC2Mixin, TestCase):
-    file_name = 'FC-2-15x15-10X.dcm'
+    file_name = "FC-2-15x15-10X.dcm"
     field_size_x_mm = 149.2
     field_size_y_mm = 149.2
     field_epid_offset_x_mm = 0.1
@@ -501,7 +571,7 @@ class FC215x15_10X(FC2Mixin, TestCase):
 
 
 class FC215x15_10FFF(FC2Mixin, TestCase):
-    file_name = 'FC-2-15x15-10XFFF.dcm'
+    file_name = "FC-2-15x15-10XFFF.dcm"
     fwxm = 30
     field_size_x_mm = 149.5
     field_size_y_mm = 149.6
@@ -512,7 +582,7 @@ class FC215x15_10FFF(FC2Mixin, TestCase):
 
 
 class FC2Yoda(FC2Mixin, TestCase):
-    file_name = 'FC-2-Yoda.dcm'
+    file_name = "FC-2-Yoda.dcm"
     field_size_y_mm = 148.2
     field_size_x_mm = 149.2
     field_epid_offset_y_mm = -1
@@ -522,7 +592,7 @@ class FC2Yoda(FC2Mixin, TestCase):
 
 
 class FC2Perfect(FC2Mixin, TestCase):
-    file_name = 'fc2-perfect.dcm'
+    file_name = "fc2-perfect.dcm"
     field_size_y_mm = 120.3
     field_size_x_mm = 120.3
     field_epid_offset_y_mm = 0
@@ -532,7 +602,7 @@ class FC2Perfect(FC2Mixin, TestCase):
 
 
 class FC2FieldDown1mm(FC2Mixin, TestCase):
-    file_name = 'fc2-down1mm.dcm'
+    file_name = "fc2-down1mm.dcm"
     field_size_y_mm = 120.3
     field_size_x_mm = 120.3
     field_epid_offset_y_mm = -1.0
@@ -542,7 +612,7 @@ class FC2FieldDown1mm(FC2Mixin, TestCase):
 
 
 class FC2BBDownRight1mm(FC2Mixin, TestCase):
-    file_name = 'fc2-bbdownright1mm.dcm'
+    file_name = "fc2-bbdownright1mm.dcm"
     field_size_y_mm = 120.3
     field_size_x_mm = 120.3
     field_epid_offset_y_mm = 0
@@ -553,7 +623,7 @@ class FC2BBDownRight1mm(FC2Mixin, TestCase):
 
 class IMTLRadMixin(FC2Mixin):
     klass = IMTLRad
-    dir_path = ['planar_imaging', 'IMT_L-Rad']
+    dir_path = ["planar_imaging", "IMT_L-Rad"]
 
 
 class IMTLRadDemo(IMTLRadMixin, TestCase):
@@ -569,7 +639,7 @@ class IMTLRadDemo(IMTLRadMixin, TestCase):
 
 
 class IMTLRad21x21(IMTLRadMixin, TestCase):
-    file_name = 'RTIMAGE_11_1.dcm'
+    file_name = "RTIMAGE_11_1.dcm"
     field_size_y_mm = 210.8
     field_size_x_mm = 210.5
     field_epid_offset_x_mm = 1.3
@@ -579,7 +649,7 @@ class IMTLRad21x21(IMTLRadMixin, TestCase):
 
 
 class IMTLRad21x21_2(IMTLRadMixin, TestCase):
-    file_name = 'RTIMAGE_11_2.dcm'
+    file_name = "RTIMAGE_11_2.dcm"
     field_size_y_mm = 210.9
     field_size_x_mm = 210.7
     field_epid_offset_x_mm = 1.4
@@ -589,7 +659,7 @@ class IMTLRad21x21_2(IMTLRadMixin, TestCase):
 
 
 class IMTLRadPerfect(IMTLRadMixin, TestCase):
-    file_name = 'perfect_imt.dcm'
+    file_name = "perfect_imt.dcm"
     field_size_y_mm = 150
     field_size_x_mm = 150
     field_epid_offset_x_mm = 0
@@ -599,7 +669,7 @@ class IMTLRadPerfect(IMTLRadMixin, TestCase):
 
 
 class IMTLRadOffset(IMTLRadMixin, TestCase):
-    file_name = 'offset_imt.dcm'
+    file_name = "offset_imt.dcm"
     field_size_y_mm = 150
     field_size_x_mm = 150
     field_epid_offset_x_mm = 0
@@ -610,7 +680,7 @@ class IMTLRadOffset(IMTLRadMixin, TestCase):
 
 class SNCFSQAMixin(FC2Mixin):
     klass = SNCFSQA
-    dir_path = ['planar_imaging', 'SNC FSQA']
+    dir_path = ["planar_imaging", "SNC FSQA"]
 
 
 class SNCFSQADemo(SNCFSQAMixin, TestCase):
@@ -627,7 +697,7 @@ class SNCFSQADemo(SNCFSQAMixin, TestCase):
 
 
 class SNCFSQA10x10(SNCFSQAMixin, TestCase):
-    file_name = '6x_FSQA_10x10.dcm'
+    file_name = "6x_FSQA_10x10.dcm"
     field_size_y_mm = 99.6
     field_size_x_mm = 100.2
     field_epid_offset_x_mm = 0.7